pyproject.toml
[project]
name = "learning-path-agent"
version = "0.1.0"
description = "AI-powered learning path customization service for engineers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.123.0",
    "langchain>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "mangum>=0.19.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.5.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
